fix(ptp): correct logic in GM ptpconfig lifecycle tests
